Add files via upload
This repository contains the necessary code to conduct analysis and reproduce the images from XX. 
The precalculated results are stored in the data folder. The required trajectories are freely available at XX for all analyses except the water analysis. 
Simulations for the water analysis are available upon request due to the large storage requirements